# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Project 2 API",
    "description": "Books and Authors API for Project 2",
    "version": "1.0.0"
  },
  "host": "cse-341-project2-1dsv.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
    "/authors": {
      "get": {
        "description": "Get all authors",
        "produces": ["application/json"],
        "responses": {"200": {"description": "OK"}, "500": {"description": "Internal Server Error"}}
      },
      "post": {
        "description": "Create an author",
        "parameters": [{"name": "body", "in": "body", "schema": {"$ref": "#/definitions/Author"}}],
        "responses": {"201": {"description": "Created"}, "400": {"description": "Bad Request"}, "500": {"description": "Internal Server Error"}}
      }
    },
    "/authors/{id}": {
      "get": {
        "description": "Get author by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {"200": {"description": "OK"}, "400": {"description": "Bad Request"}, "404": {"description": "Not Found"}, "500": {"description": "Internal Server Error"}}
      },
      "put": {
        "description": "Update author by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}, {"name": "body", "in": "body", "schema": {"$ref": "#/definitions/Author"}}],
        "responses": {"204": {"description": "No Content"}, "400": {"description": "Bad Request"}, "500": {"description": "Internal Server Error"}}
      },
      "delete": {
        "description": "Delete author by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {"204": {"description": "No Content"}, "400": {"description": "Bad Request"}, "404": {"description": "Not Found"}, "500": {"description": "Internal Server Error"}}
      }
    },
    "/books": {
      "get": {
        "description": "Get all books",
        "produces": ["application/json"],
        "responses": {"200": {"description": "OK"}, "500": {"description": "Internal Server Error"}}
      },
      "post": {
        "description": "Create a book",
        "parameters": [{"name": "body", "in": "body", "schema": {"$ref": "#/definitions/Book"}}],
        "responses": {"201": {"description": "Created"}, "400": {"description": "Bad Request"}, "500": {"description": "Internal Server Error"}}
      }
    },
    "/books/{id}": {
      "get": {
        "description": "Get book by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {"200": {"description": "OK"}, "400": {"description": "Bad Request"}, "404": {"description": "Not Found"}, "500": {"description": "Internal Server Error"}}
      },
      "put": {
        "description": "Update book by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}, {"name": "body", "in": "body", "schema": {"$ref": "#/definitions/Book"}}],
        "responses": {"204": {"description": "No Content"}, "400": {"description": "Bad Request"}, "500": {"description": "Internal Server Error"}}
      },
      "delete": {
        "description": "Delete book by id",
        "parameters": [{"name": "id", "in": "path", "required": true, "type": "string"}],
        "responses": {"204": {"description": "No Content"}, "400": {"description": "Bad Request"}, "404": {"description": "Not Found"}, "500": {"description": "Internal Server Error"}}
      }
    }
  },
  "definitions": {
    "Author": {
      "type": "object",
      "properties": {
        "name": {"type": "string"},
        "bio": {"type": "string"},
        "birthDate": {"type": "string", "format": "date"},
        "nationality": {"type": "string"}
      },
      "required": ["name"]
    },
    "Book": {
      "type": "object",
      "properties": {
        "title": {"type": "string"},
        "authorId": {"type": "string"},
        "isbn": {"type": "string"},
        "publishedDate": {"type": "string", "format": "date"},
        "pages": {"type": "integer"},
        "genre": {"type": "string"},
        "summary": {"type": "string"},
        "rating": {"type": "number"}
      },
      "required": ["title", "authorId", "isbn", "publishedDate", "pages", "genre", "rating"]
    }
  }
}